backend/requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
diff-match-patch==20241021
Django==6.0.1
django-cors-headers==4.7.0
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.9.0
python-dotenv==1.2.1
reportlab==4.4.9
sqlparse==0.5.3
tablib==3.8.0
whitenoise==6.11.0
dj-database-url>=2.1
openpyxl==3.1.2
requests==2.32.5

coverage==7.8.0
